Remoived size_t/int comparison

+++ encoder/basisu_enc.cpp
@@ -1667,7 +1667,7 @@ namespace basisu
 		WORD const numGroups = GetActiveProcessorGroupCount();
 		if (numGroups > 1) {
 			GROUP_AFFINITY affinity;
-			for (unsigned index = 0; index < threads; ) {
+			for (size_t index = 0; index < threads;) {
 				for (WORD group = 0; group < numGroups; ++group) {
 					if (DWORD  groupCpus = GetMaximumProcessorCount(group)) {
 						for (; groupCpus > 0; --groupCpus) {
